make everything in one file, because vercel is a crybaby
